backend/requirements-dev.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.21
redis==5.0.0
pyjwt==2.8.0
bcrypt==4.0.1
gunicorn==21.2.0
